[reflink] support hardlinks
a suffix made of tid is now used until the changes are commited. As such, when linking the file, we should link to the suffixed file.

The suffix is currently internal only to rpmPackageFilesInstall so I am duplicating the work here. A better approach would be to pass *fp* directly to the plugin as the latter has all the context needed.